// src/features/EditableProfileCard/model/consts/consts.ts
export enum ProfileValidationErrors {
  INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
  INCORRECT_USER_AGE = 'INCORRECT_USER_AGE',
  INCORRECT_USER_CURRENCY = 'INCORRECT_USER_CURRENCY',
  INCORRECT_USER_COUNTRY = 'INCORRECT_USER_COUNTRY',
  INCORRECT_USER_CITY = 'INCORRECT_USER_CITY',
  INCORRECT_USER_USERNAME = 'INCORRECT_USER_USERNAME',
  INCORRECT_USER_AVATAR = 'INCORRECT_USER_AVATAR',
  NO_DATA = 'NO_DATA',
  SERVER_ERROR = 'SERVER_ERROR',
}